// src/app/models/user.model.ts
export interface User {
  id: number;
  name: string;
  nickname: string;
  email: string;
  phone: string;
  address: string;
  address2: string;
  address3: string;
  address4: string;
  postalCode: string;
  state: string;
  municipality: string;
  houseDescription: string;
  countryName: string;
}